l3arn-platform/apps/api/requirements.txt
# FastAPI Backend — L3ARN Platform
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
httpx>=0.28.0,<1.0.0
structlog>=24.4.0,<25.0.0
supabase>=2.11.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-multipart>=0.0.17,<1.0.0

# Dev / Test
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
ruff>=0.8.0,<1.0.0
mypy>=1.13.0,<2.0.0
